// front/src/components/page-builder/Render/types.ts
export type ElementTypes =
  // base elements
  | 'html'
  | 'text'
  | 'button'
  | 'hr'
  | 'header'
  | 'TEXTBOX'
  | 'grid'
  | 'form'
  | 'TEXTNODE'
  | 'row'
  | 'col'
  | 'Cell' // alias to col
  | 'image'
  | 'Content'
  | 'tabs'
  | 'tab'
  | 'sidemenu'

  // sliders
  | 'swiper-container'
  | 'slider'
  | 'Slide'
  | 'CAROUSEL'
  | 'ProductSlider'
  | 'PostSlider'

  // ads
  | 'advertises'
  | 'addadvertises'
  | 'adbanner'

  // gomrok
  | 'currency'
  | 'calculation'
  | 'packinginvoice'
  | 'packing'
  | 'invoice'
  | 'tsc'
  | 'chatgpt'
  | 'flights'
  | 'backButton'

  //
  | 'fastcheckout'
  | 'ProductElement'

  // pagination
  | 'loadmore'
  | 'pagination'
  | 'description'
  | 'navigation'

  // steps
  | 'conditionsteps'
  | 'conditionstep'

  // search
  | 'searchbar'
  | 'searchbarinitem'
  | 'searchbaritem'
  | 'searchbartsc';

// custom